+++ src/N2kMaretron.h
@@ -1,7 +1,7 @@
 /*
  * N2kMaretron.h
  * 
- * Copyright (c) 2019-2023 Vassilis Bourdakis,
+ * Copyright (c) 2019-2024 Vassilis Bourdakis,
  *                        Timo Lappalainen, Kave Oy, www.kave.fi, 
  * 
  * Permission is hereby granted, free of charge, to any person obtaining a
@@ -45,6 +45,7 @@
 #define _N2kMaretron_H_
 
 #include "N2kMsg.h"
+#include "N2kTypes.h"
 #include <stdint.h>