+++ src/Seasmart.cpp
@@ -1,7 +1,7 @@
 /*
   The MIT License
 
-  Copyright (c) 2017-2023 Thomas Sarlandie thomas@sarlandie.net
+  Copyright (c) 2017-2024 Thomas Sarlandie thomas@sarlandie.net
 
   Permission is hereby granted, free of charge, to any person obtaining a copy
   of this software and associated documentation files (the "Software"), to deal